requirements-dev.txt
-r requirements.txt
pytest==3.0.2
pytest-flask==0.10.0
factory-boy==2.7.0
factory-boy-peewee==0.0.4
flex==5.8.0
